refactor(peer-collab): give mobile and peer clients separate RPC allowlists
Mobile and peer connections shared one 248-method allowlist, but they are
not the same trust layer: mobile is the same user's own device extending
their runtime, while a peer is a different person's desktop admitted only
to share a terminal. Sharing the list meant a paired peer could reach the
host's browser controls, account list, AI vault sessions, and file and git
methods — none of which terminal sharing needs.

Each scope now checks its own list. The peer list holds the seven methods
its client actually calls, so the collaboration surface is what the feature
needs and nothing more, and presence and subscriber listing move off the
mobile list since only the 1:N layer uses them. Adding a method now forces
picking a layer instead of silently widening both.

The existing per-terminal grant guards stay as they were; this is a second
layer in front of them, not a replacement.
