+++ src/main/runtime/mobile-rpc-allowlist.test.ts
@@ -95,8 +95,13 @@ function mobileRpcMethods(): string[] {
 }
 
 function mobileRpcAllowlist(): Set<string> {
-  const source = readFileSync(join(process.cwd(), 'src/main/runtime/runtime-rpc.ts'), 'utf8')
-  const allowlist = source.match(/const MOBILE_RPC_METHOD_ALLOWLIST = new Set\(\[([\s\S]*?)\]\)/)
+  const source = readFileSync(
+    join(process.cwd(), 'src/main/runtime/rpc/scoped-rpc-method-allowlists.ts'),
+    'utf8'
+  )
+  const allowlist = source.match(
+    /export const MOBILE_RPC_METHOD_ALLOWLIST = new Set\(\[([\s\S]*?)\]\)/
+  )
   if (!allowlist) {
     throw new Error('MOBILE_RPC_METHOD_ALLOWLIST not found')
   }
