+++ src/main/runtime/rpc/scoped-rpc-method-allowlists.ts
@@ -1,0 +1,279 @@
+// Why: mobile is a 1:1 same-user extension of this runtime; a peer is a
+// different person's desktop joined for terminal sharing. Distinct trust
+// models get distinct allowlists so adding a method forces picking a layer.
+export const MOBILE_RPC_METHOD_ALLOWLIST = new Set([
+  'accounts.list',
+  'accounts.consumeCodexResetCredit',
+  'accounts.selectClaude',
+  'accounts.selectCodex',
+  'accounts.selectCodexForTarget',
+  'accounts.subscribe',
+  'accounts.unsubscribe',
+  'aiVault.listSessions',
+  'aiVault.prepareSessionResume',
+  'browser.back',
+  'browser.dialogAccept',
+  'browser.dialogDismiss',
+  'browser.forward',
+  'browser.goto',
+  'browser.keyboardInsertText',
+  'browser.keypress',
+  'browser.mouseDown',
+  'browser.mouseClick',
+  'browser.mouseMove',
+  'browser.mouseUp',
+  'browser.mouseWheel',
+  'browser.reload',
+  'browser.screencast',
+  'browser.screencast.unsubscribe',
+  'browser.tabCreate',
+  'browser.viewport',
+  'clipboard.abortImageUpload',
+  'clipboard.appendImageUploadChunk',
+  'clipboard.commitImageUpload',
+  'clipboard.saveImageAsTempFile',
+  'clipboard.startImageUpload',
+  'diagnostics.memory',
+  'files.browseServerDir',
+  'files.createFile',
+  'files.list',
+  'files.open',
+  'files.openDiff',
+  'files.read',
+  'files.readChunk',
+  'files.readDir',
+  'files.readPreview',
+  'files.readTerminalArtifact',
+  'files.readTerminalArtifactPreview',
+  'files.resolveTerminalPath',
+  'files.searchPaths',
+  'files.writeTerminalArtifact',
+  'folderWorkspace.list',
+  'git.abortMerge',
+  'git.abortRebase',
+  'git.bulkStage',
+  'git.bulkUnstage',
+  'git.branchCompare',
+  'git.branchDiff',
+  'git.cancelGenerateCommitMessage',
+  'git.cancelGeneratePullRequestFields',
+  'git.checkout',
+  'git.commit',
+  'git.commitCompare',
+  'git.commitDiff',
+  'git.discard',
+  'git.discoverCommitMessageModels',
+  'git.diff',
+  'git.fetch',
+  'git.forkSync',
+  'git.fastForward',
+  'git.generateCommitMessage',
+  'git.generatePullRequestFields',
+  'git.history',
+  'git.localBranches',
+  'git.pull',
+  'git.push',
+  'git.rebaseFromBase',
+  'git.stage',
+  'git.status',
+  'git.unstage',
+  'git.upstreamStatus',
+  'github.createIssue',
+  'github.addIssueComment',
+  'github.addPRReviewComment',
+  'github.addPRReviewCommentReply',
+  'github.countWorkItems',
+  'github.listAssignableUsers',
+  'github.listLabels',
+  'github.listWorkItems',
+  'github.mergePR',
+  'github.setPRAutoMerge',
+  'github.requestPRReviewers',
+  'github.removePRReviewers',
+  'github.project.listAccessible',
+  'github.project.listAssignableUsersBySlug',
+  'github.project.listIssueTypesBySlug',
+  'github.project.listLabelsBySlug',
+  'github.project.listViews',
+  'github.project.resolveRef',
+  'github.project.addIssueCommentBySlug',
+  'github.project.updateIssueCommentBySlug',
+  'github.project.deleteIssueCommentBySlug',
+  'github.project.clearItemField',
+  'github.project.updateIssueBySlug',
+  'github.project.updateIssueTypeBySlug',
+  'github.project.updateItemField',
+  'github.project.updatePullRequestBySlug',
+  'github.project.viewTable',
+  'github.project.workItemDetailsBySlug',
+  'github.prForBranch',
+  'github.prFileContents',
+  'github.prChecks',
+  'github.prCheckDetails',
+  'github.rerunPRChecks',
+  'github.resolveReviewThread',
+  'github.setPRFileViewed',
+  'github.updateIssue',
+  'github.updatePR',
+  'github.updatePRTitle',
+  'github.updatePRState',
+  'github.repoSlug',
+  'github.workItem',
+  // Cross-repo lookup: lets the mobile Smart picker resolve a pasted github.com URL for a different repo.
+  'github.workItemByOwnerRepo',
+  'github.workItemDetails',
+  'gitlab.createIssue',
+  'gitlab.addIssueComment',
+  'gitlab.addMRComment',
+  'gitlab.listWorkItems',
+  // Mobile Smart picker: resolve a pasted GitLab URL to an exact issue/MR (MR listing reuses gitlab.listWorkItems).
+  'gitlab.workItemByPath',
+  'gitlab.mergeMR',
+  'gitlab.resolveMRDiscussion',
+  'gitlab.todos',
+  'gitlab.updateIssue',
+  'gitlab.updateMR',
+  'gitlab.updateMRState',
+  'gitlab.workItemDetails',
+  'host.gitBash.isAvailable',
+  'host.platform',
+  'host.pwsh.isAvailable',
+  'host.wsl.isAvailable',
+  'host.wsl.listDistros',
+  'hostedReview.create',
+  'hostedReview.forBranch',
+  'hostedReview.getCreationEligibility',
+  'linear.getCustomView',
+  'linear.getIssue',
+  'linear.getProject',
+  'linear.agentSearchIssues',
+  'linear.issueContext',
+  'linear.resolveCurrentIssue',
+  'linear.addIssueComment',
+  'linear.connect',
+  'linear.createIssue',
+  'linear.createProject',
+  'linear.issueComments',
+  'linear.listCustomViewIssues',
+  'linear.listCustomViewProjects',
+  'linear.listCustomViews',
+  'linear.listIssues',
+  'linear.mcpListIssues',
+  'linear.listProjectIssues',
+  'linear.listProjects',
+  'linear.teamLabels',
+  'linear.teamMembers',
+  'linear.listTeams',
+  'linear.searchIssues',
+  'linear.selectWorkspace',
+  'linear.status',
+  'linear.teamStates',
+  'linear.updateIssue',
+  'markdown.readTab',
+  'markdown.saveTab',
+  'notifications.getMissedSince',
+  'notifications.subscribe',
+  'notifications.unsubscribe',
+  'pairing.getEndpoints',
+  'pairing.provisionRelay',
+  'preflight.check',
+  'preflight.detectAgents',
+  'preflight.detectRemoteAgents',
+  'projectGroup.list',
+  'repo.baseRefDefault',
+  'repo.gitAvailable',
+  'repo.hooks',
+  'repo.list',
+  'repo.saveSparsePreset',
+  'repo.searchRefs',
+  'repo.sparsePresets',
+  'repo.update',
+  'runtime.clientEvents.subscribe',
+  'runtime.clientEvents.unsubscribe',
+  'session.tabs.activate',
+  'session.tabs.close',
+  'session.tabs.closeLifecycle',
+  'session.tabs.createTerminal',
+  'session.tabs.list',
+  'session.tabs.listAll',
+  'session.tabs.move',
+  'session.tabs.subscribe',
+  'session.tabs.subscribeAll',
+  'session.tabs.unsubscribe',
+  'session.tabs.unsubscribeAll',
+  'nativeChat.readSession',
+  'nativeChat.subscribe',
+  'nativeChat.unsubscribe',
+  'settings.get',
+  'settings.getTerminalQuickCommands',
+  'settings.update',
+  'settings.updateTerminalQuickCommands',
+  'ssh.connect',
+  'ssh.getState',
+  'ssh.listRemovedTargetLabels',
+  'ssh.listTargets',
+  'ssh.listTargetSummaries',
+  'speech.dictation.cancel',
+  'speech.dictation.chunk',
+  'speech.dictation.finish',
+  'speech.dictation.setup',
+  'speech.dictation.start',
+  'speech.models.delete',
+  'speech.models.download',
+  'speech.models.list',
+  'stats.summary',
+  'status.get',
+  'agentTeams.prepareLaunch',
+  'agentTeams.tmuxCompat',
+  'terminal.clearBuffer',
+  'terminal.close',
+  'terminal.closeTab',
+  'terminal.create',
+  'terminal.createAgentSession',
+  'terminal.ensureAgentSession',
+  'terminal.focus',
+  'terminal.agentStatus',
+  'terminal.adoptOrphans',
+  'terminal.getAutoRestoreFit',
+  'terminal.isRunningAgent',
+  'terminal.list',
+  'terminal.multiplex',
+  'terminal.read',
+  'terminal.rename',
+  'terminal.send',
+  'terminal.setAutoRestoreFit',
+  'terminal.setDisplayMode',
+  'terminal.subscribe',
+  'terminal.unsubscribe',
+  'terminal.updateViewport',
+  'terminal.wait',
+  'ui.get',
+  'ui.recordFeatureInteraction',
+  'ui.set',
+  'worktree.activate',
+  'worktree.create',
+  'worktree.forceDeleteBranch',
+  'worktree.prefetchCreateBase',
+  'worktree.ps',
+  'worktree.show',
+  'worktree.resolveMrBase',
+  'worktree.resolvePrBase',
+  'worktree.rm',
+  'worktree.set',
+  'worktree.sleep'
+])
+
+// Why: the only RPC surface src/main/runtime/peer-client-*.ts actually calls —
+// terminal.send/resize ride the binary stream, not RPC. Grown only when a new
+// peer-client call site needs it (see peer-client-service.ts, peer-client-terminal-streams.ts,
+// peer-client-presence-streams.ts); the e2ee handshake and reconnect loop don't
+// dispatch any RPC method, so they need no entries here.
+export const PEER_RPC_METHOD_ALLOWLIST = new Set([
+  'terminal.list',
+  'terminal.subscribe',
+  'terminal.unsubscribe',
+  'terminal.listSubscribers',
+  'terminal.presence.subscribe',
+  'terminal.presence.send',
+  'terminal.presence.unsubscribe'
+])
