+++ src/main/runtime/runtime-rpc.ts
@@ -28,6 +28,10 @@ import {
   type MobileSocketTransportMetadata
 } from './rpc/mobile-socket-wiring'
 import { PeerConnectionRegistry } from './peer-connection-registry'
+import {
+  MOBILE_RPC_METHOD_ALLOWLIST,
+  PEER_RPC_METHOD_ALLOWLIST
+} from './rpc/scoped-rpc-method-allowlists'
 import { PEER_DUPLICATE_CONNECTION_CLOSE_CODE } from '../../shared/peer-connection-close-codes'
 import type { PairingRelay } from '../../shared/mobile-relay-pairing-offer'
 import type { MobilePairingConnectionMode } from '../../shared/mobile-pairing-connection-mode'
@@ -165,272 +169,6 @@ function webClientPathForEndpoint(pathname: string): string {
   }
   return `${pathname.replace(/\/$/, '')}/web-index.html`
 }
-
-const MOBILE_RPC_METHOD_ALLOWLIST = new Set([
-  'accounts.list',
-  'accounts.consumeCodexResetCredit',
-  'accounts.selectClaude',
-  'accounts.selectCodex',
-  'accounts.selectCodexForTarget',
-  'accounts.subscribe',
-  'accounts.unsubscribe',
-  'aiVault.listSessions',
-  'aiVault.prepareSessionResume',
-  'browser.back',
-  'browser.dialogAccept',
-  'browser.dialogDismiss',
-  'browser.forward',
-  'browser.goto',
-  'browser.keyboardInsertText',
-  'browser.keypress',
-  'browser.mouseDown',
-  'browser.mouseClick',
-  'browser.mouseMove',
-  'browser.mouseUp',
-  'browser.mouseWheel',
-  'browser.reload',
-  'browser.screencast',
-  'browser.screencast.unsubscribe',
-  'browser.tabCreate',
-  'browser.viewport',
-  'clipboard.abortImageUpload',
-  'clipboard.appendImageUploadChunk',
-  'clipboard.commitImageUpload',
-  'clipboard.saveImageAsTempFile',
-  'clipboard.startImageUpload',
-  'diagnostics.memory',
-  'files.browseServerDir',
-  'files.createFile',
-  'files.list',
-  'files.open',
-  'files.openDiff',
-  'files.read',
-  'files.readChunk',
-  'files.readDir',
-  'files.readPreview',
-  'files.readTerminalArtifact',
-  'files.readTerminalArtifactPreview',
-  'files.resolveTerminalPath',
-  'files.searchPaths',
-  'files.writeTerminalArtifact',
-  'folderWorkspace.list',
-  'git.abortMerge',
-  'git.abortRebase',
-  'git.bulkStage',
-  'git.bulkUnstage',
-  'git.branchCompare',
-  'git.branchDiff',
-  'git.cancelGenerateCommitMessage',
-  'git.cancelGeneratePullRequestFields',
-  'git.checkout',
-  'git.commit',
-  'git.commitCompare',
-  'git.commitDiff',
-  'git.discard',
-  'git.discoverCommitMessageModels',
-  'git.diff',
-  'git.fetch',
-  'git.forkSync',
-  'git.fastForward',
-  'git.generateCommitMessage',
-  'git.generatePullRequestFields',
-  'git.history',
-  'git.localBranches',
-  'git.pull',
-  'git.push',
-  'git.rebaseFromBase',
-  'git.stage',
-  'git.status',
-  'git.unstage',
-  'git.upstreamStatus',
-  'github.createIssue',
-  'github.addIssueComment',
-  'github.addPRReviewComment',
-  'github.addPRReviewCommentReply',
-  'github.countWorkItems',
-  'github.listAssignableUsers',
-  'github.listLabels',
-  'github.listWorkItems',
-  'github.mergePR',
-  'github.setPRAutoMerge',
-  'github.requestPRReviewers',
-  'github.removePRReviewers',
-  'github.project.listAccessible',
-  'github.project.listAssignableUsersBySlug',
-  'github.project.listIssueTypesBySlug',
-  'github.project.listLabelsBySlug',
-  'github.project.listViews',
-  'github.project.resolveRef',
-  'github.project.addIssueCommentBySlug',
-  'github.project.updateIssueCommentBySlug',
-  'github.project.deleteIssueCommentBySlug',
-  'github.project.clearItemField',
-  'github.project.updateIssueBySlug',
-  'github.project.updateIssueTypeBySlug',
-  'github.project.updateItemField',
-  'github.project.updatePullRequestBySlug',
-  'github.project.viewTable',
-  'github.project.workItemDetailsBySlug',
-  'github.prForBranch',
-  'github.prFileContents',
-  'github.prChecks',
-  'github.prCheckDetails',
-  'github.rerunPRChecks',
-  'github.resolveReviewThread',
-  'github.setPRFileViewed',
-  'github.updateIssue',
-  'github.updatePR',
-  'github.updatePRTitle',
-  'github.updatePRState',
-  'github.repoSlug',
-  'github.workItem',
-  // Cross-repo lookup: lets the mobile Smart picker resolve a pasted github.com URL for a different repo.
-  'github.workItemByOwnerRepo',
-  'github.workItemDetails',
-  'gitlab.createIssue',
-  'gitlab.addIssueComment',
-  'gitlab.addMRComment',
-  'gitlab.listWorkItems',
-  // Mobile Smart picker: resolve a pasted GitLab URL to an exact issue/MR (MR listing reuses gitlab.listWorkItems).
-  'gitlab.workItemByPath',
-  'gitlab.mergeMR',
-  'gitlab.resolveMRDiscussion',
-  'gitlab.todos',
-  'gitlab.updateIssue',
-  'gitlab.updateMR',
-  'gitlab.updateMRState',
-  'gitlab.workItemDetails',
-  'host.gitBash.isAvailable',
-  'host.platform',
-  'host.pwsh.isAvailable',
-  'host.wsl.isAvailable',
-  'host.wsl.listDistros',
-  'hostedReview.create',
-  'hostedReview.forBranch',
-  'hostedReview.getCreationEligibility',
-  'linear.getCustomView',
-  'linear.getIssue',
-  'linear.getProject',
-  'linear.agentSearchIssues',
-  'linear.issueContext',
-  'linear.resolveCurrentIssue',
-  'linear.addIssueComment',
-  'linear.connect',
-  'linear.createIssue',
-  'linear.createProject',
-  'linear.issueComments',
-  'linear.listCustomViewIssues',
-  'linear.listCustomViewProjects',
-  'linear.listCustomViews',
-  'linear.listIssues',
-  'linear.mcpListIssues',
-  'linear.listProjectIssues',
-  'linear.listProjects',
-  'linear.teamLabels',
-  'linear.teamMembers',
-  'linear.listTeams',
-  'linear.searchIssues',
-  'linear.selectWorkspace',
-  'linear.status',
-  'linear.teamStates',
-  'linear.updateIssue',
-  'markdown.readTab',
-  'markdown.saveTab',
-  'notifications.getMissedSince',
-  'notifications.subscribe',
-  'notifications.unsubscribe',
-  'pairing.getEndpoints',
-  'pairing.provisionRelay',
-  'preflight.check',
-  'preflight.detectAgents',
-  'preflight.detectRemoteAgents',
-  'projectGroup.list',
-  'repo.baseRefDefault',
-  'repo.gitAvailable',
-  'repo.hooks',
-  'repo.list',
-  'repo.saveSparsePreset',
-  'repo.searchRefs',
-  'repo.sparsePresets',
-  'repo.update',
-  'runtime.clientEvents.subscribe',
-  'runtime.clientEvents.unsubscribe',
-  'session.tabs.activate',
-  'session.tabs.close',
-  'session.tabs.closeLifecycle',
-  'session.tabs.createTerminal',
-  'session.tabs.list',
-  'session.tabs.listAll',
-  'session.tabs.move',
-  'session.tabs.subscribe',
-  'session.tabs.subscribeAll',
-  'session.tabs.unsubscribe',
-  'session.tabs.unsubscribeAll',
-  'nativeChat.readSession',
-  'nativeChat.subscribe',
-  'nativeChat.unsubscribe',
-  'settings.get',
-  'settings.getTerminalQuickCommands',
-  'settings.update',
-  'settings.updateTerminalQuickCommands',
-  'ssh.connect',
-  'ssh.getState',
-  'ssh.listRemovedTargetLabels',
-  'ssh.listTargets',
-  'ssh.listTargetSummaries',
-  'speech.dictation.cancel',
-  'speech.dictation.chunk',
-  'speech.dictation.finish',
-  'speech.dictation.setup',
-  'speech.dictation.start',
-  'speech.models.delete',
-  'speech.models.download',
-  'speech.models.list',
-  'stats.summary',
-  'status.get',
-  'agentTeams.prepareLaunch',
-  'agentTeams.tmuxCompat',
-  'terminal.clearBuffer',
-  'terminal.close',
-  'terminal.closeTab',
-  'terminal.create',
-  'terminal.createAgentSession',
-  'terminal.ensureAgentSession',
-  'terminal.focus',
-  'terminal.agentStatus',
-  'terminal.adoptOrphans',
-  'terminal.getAutoRestoreFit',
-  'terminal.isRunningAgent',
-  'terminal.list',
-  'terminal.listSubscribers',
-  'terminal.multiplex',
-  'terminal.presence.subscribe',
-  'terminal.presence.send',
-  'terminal.presence.unsubscribe',
-  'terminal.read',
-  'terminal.rename',
-  'terminal.send',
-  'terminal.setAutoRestoreFit',
-  'terminal.setDisplayMode',
-  'terminal.subscribe',
-  'terminal.unsubscribe',
-  'terminal.updateViewport',
-  'terminal.wait',
-  'ui.get',
-  'ui.recordFeatureInteraction',
-  'ui.set',
-  'worktree.activate',
-  'worktree.create',
-  'worktree.forceDeleteBranch',
-  'worktree.prefetchCreateBase',
-  'worktree.ps',
-  'worktree.show',
-  'worktree.resolveMrBase',
-  'worktree.resolvePrBase',
-  'worktree.rm',
-  'worktree.set',
-  'worktree.sleep'
-])
 
 // Why: 'ask' is metered separately from 'wait' — same keepalive/abort wiring, its own sub-cap.
 type LongPollClass = 'ask' | 'wait'
@@ -1520,13 +1258,17 @@ export class OrcaRuntimeRpcServer {
       reply(JSON.stringify(this.buildError(request.id, 'unauthorized', 'Invalid device token')))
       return
     }
-    // Why: peer desktops get the same restricted RPC surface as mobile, not the
-    // unrestricted 'runtime' scope — a LAN peer must not gain broader access
-    // than the terminal-sharing feature it was paired for.
-    if (
-      (device.scope === 'mobile' || device.scope === 'peer') &&
-      !MOBILE_RPC_METHOD_ALLOWLIST.has(request.method)
-    ) {
+    // Why: mobile and peer are different trust models — mobile is the same
+    // user's own device, a peer is a different person's desktop admitted only
+    // for terminal sharing — so each scope is checked against its own
+    // allowlist instead of a shared one that would over-grant the peer.
+    const scopedAllowlist =
+      device.scope === 'mobile'
+        ? MOBILE_RPC_METHOD_ALLOWLIST
+        : device.scope === 'peer'
+          ? PEER_RPC_METHOD_ALLOWLIST
+          : null
+    if (scopedAllowlist && !scopedAllowlist.has(request.method)) {
       reply(
         JSON.stringify(
           this.buildError(
