+++ src/renderer/src/components/settings/PeerCollabSettingsPane.tsx
@@ -8,7 +8,7 @@ import {
   type HostTerminalOption
 } from './PeerCollabConnectedClientsSection'
 import { PeerCollabClientConnectSection } from './PeerCollabClientConnectSection'
-import { usePeerCollabSavedPairing } from './use-peer-collab-saved-pairing'
+import { usePeerCollabClientConnection } from './use-peer-collab-client-connection'
 import type { PairedDevice } from './MobilePairedDevicesSection'
 import {
   selectRefreshedNetworkAddress,
@@ -16,13 +16,7 @@ import {
 } from './mobile-network-interface-selection'
 import { useMountedRef } from '@/hooks/useMountedRef'
 import { translate } from '@/i18n/i18n'
-import type { RuntimeTerminalListResult } from '../../../../shared/runtime-types'
-import type { PeerClientStatus } from '../../../../shared/peer-client-status'
-import {
-  RemoteTerminalDialog,
-  type RemoteTerminalDialogTarget
-} from '@/components/peer-collab/RemoteTerminalDialog'
-import { isTerminalListResult } from './peer-collab-terminal-list-guard'
+import { RemoteTerminalDialog } from '@/components/peer-collab/RemoteTerminalDialog'
 export { getPeerCollabSettingsPaneSearchEntries } from './peer-collab-settings-search'
 
 // Why: no push channel exists for peer connection changes yet, so the
@@ -42,29 +36,26 @@ export function PeerCollabSettingsPane(): React.JSX.Element {
   const [devices, setDevices] = useState<PairedDevice[]>([])
   const [connectedClients, setConnectedClients] = useState<ConnectedPeerClient[]>([])
   const [exclusiveInputFloor, setExclusiveInputFloorState] = useState(false)
-  const [clientPairingCode, setClientPairingCode] = useState('')
-  const [clientDisplayName, setClientDisplayName] = useState('')
-  const [clientStatus, setClientStatus] = useState<PeerClientStatus>({
-    state: 'closed',
-    endpoint: null,
-    reconnectAttempt: 0,
-    lastErrorReason: null
-  })
-  const [clientConnectBusy, setClientConnectBusy] = useState(false)
-  const {
-    savedPairing,
-    refresh: refreshSavedPairing,
-    connectSaved,
-    forget: forgetSavedPairing
-  } = usePeerCollabSavedPairing()
-  const [hostTerminals, setHostTerminals] = useState<RuntimeTerminalListResult['terminals']>([])
-  // Why: distinct from `hostTerminals` above, which is the *remote* host's
-  // terminals seen while connected as a peer client; this is this desktop's
-  // own terminals, used for the grant picker when acting as the host.
+  // Why: distinct from `hostTerminals` in usePeerCollabClientConnection, which is the
+  // *remote* host's terminals seen while connected as a peer client; this is this
+  // desktop's own terminals, used for the grant picker when acting as the host.
   const [ownTerminals, setOwnTerminals] = useState<HostTerminalOption[]>([])
-  const [openRemoteTerminal, setOpenRemoteTerminal] = useState<RemoteTerminalDialogTarget | null>(
-    null
-  )
+  const {
+    hostTerminals,
+    clientPairingCode,
+    setClientPairingCode,
+    clientDisplayName,
+    setClientDisplayName,
+    clientStatus,
+    clientConnectBusy,
+    connectAsClient,
+    connectSavedAsClient,
+    disconnectAsClient,
+    savedPairing,
+    forgetSavedPairing,
+    openRemoteTerminal,
+    setOpenRemoteTerminal
+  } = usePeerCollabClientConnection()
   const codeCopiedResetTimerRef = useRef<number | null>(null)
   const selectedAddressRef = useRef<string | undefined>(selectedAddress)
   const mountedRef = useMountedRef()
@@ -183,74 +174,6 @@ export function PeerCollabSettingsPane(): React.JSX.Element {
     setEndpoint(null)
   }, [])
 
-  const loadHostTerminals = useCallback(async () => {
-    try {
-      const result = await window.api.peerClient.listHostTerminals()
-      if (mountedRef.current && result.ok && isTerminalListResult(result.terminals)) {
-        setHostTerminals(result.terminals.terminals)
-      }
-    } catch {
-      // Silently fail — polled again shortly while connected
-    }
-  }, [mountedRef])
-
-  const notifyClientConnectFailure = useCallback(
-    (reason: string) => {
-      if (mountedRef.current) {
-        toast.error(
-          translate(
-            'auto.components.settings.PeerCollabSettingsPane.clientConnectFailed',
-            'Failed to connect: {{reason}}',
-            { reason }
-          )
-        )
-      }
-    },
-    [mountedRef]
-  )
-
-  const runClientConnect = useCallback(
-    async (connectFn: () => Promise<{ ok: true } | { ok: false; reason: string }>) => {
-      setClientConnectBusy(true)
-      try {
-        const result = await connectFn()
-        if (!result.ok) {
-          notifyClientConnectFailure(result.reason)
-        }
-      } catch {
-        notifyClientConnectFailure('unknown_error')
-      } finally {
-        if (mountedRef.current) {
-          setClientConnectBusy(false)
-        }
-      }
-    },
-    [mountedRef, notifyClientConnectFailure]
-  )
-
-  const connectAsClient = useCallback(
-    () =>
-      runClientConnect(() =>
-        window.api.peerClient.connect({
-          pairingCode: clientPairingCode,
-          displayName: clientDisplayName
-        })
-      ),
-    [runClientConnect, clientPairingCode, clientDisplayName]
-  )
-
-  const connectSavedAsClient = useCallback(
-    () => runClientConnect(connectSaved),
-    [runClientConnect, connectSaved]
-  )
-
-  const disconnectAsClient = useCallback(async () => {
-    await window.api.peerClient.disconnect()
-    if (mountedRef.current) {
-      setHostTerminals([])
-    }
-  }, [mountedRef])
-
   const toggleExclusiveInputFloor = useCallback(async () => {
     const next = !exclusiveInputFloor
     setExclusiveInputFloorState(next)
@@ -276,17 +199,6 @@ export function PeerCollabSettingsPane(): React.JSX.Element {
         setExclusiveInputFloorState(enabled)
       }
     })
-    void window.api.peerClient.getStatus().then((status) => {
-      if (mountedRef.current) {
-        setClientStatus(status)
-      }
-    })
-    void window.api.peerClient.getDefaultDisplayName().then(({ name }) => {
-      if (mountedRef.current) {
-        setClientDisplayName(name)
-      }
-    })
-    return window.api.peerClient.onStatusChanged((status) => setClientStatus(status))
   }, [loadNetworkInterfaces, loadDevices, loadConnectedClients, loadOwnTerminals, mountedRef])
 
   useEffect(() => {
@@ -298,25 +210,6 @@ export function PeerCollabSettingsPane(): React.JSX.Element {
     const timer = window.setInterval(() => void loadOwnTerminals(), CONNECTED_CLIENTS_POLL_MS)
     return () => window.clearInterval(timer)
   }, [loadOwnTerminals])
-
-  useEffect(() => {
-    if (clientStatus.state !== 'connected') {
-      return
-    }
-    void loadHostTerminals()
-    const timer = window.setInterval(() => void loadHostTerminals(), CONNECTED_CLIENTS_POLL_MS)
-    return () => window.clearInterval(timer)
-  }, [clientStatus.state, loadHostTerminals])
-
-  useEffect(() => {
-    if (clientStatus.state === 'closed') {
-      // Why: a lost host connection leaves the stream dead — close the dialog instead of showing a frozen terminal.
-      setOpenRemoteTerminal(null)
-    } else if (clientStatus.state === 'connected') {
-      // Why: a fresh success may have just persisted a new saved pairing on the main side.
-      void refreshSavedPairing()
-    }
-  }, [clientStatus.state, refreshSavedPairing])
 
   async function revokeDevice(deviceId: string) {
     try {
